4. Формы. Загрузка файла на сервер. [#504855]

+++ src/main/java/ru/job4j/dreamjob/service/SimpleCandidateService.java
@@ -3,6 +3,7 @@ package ru.job4j.dreamjob.service;
 import net.jcip.annotations.GuardedBy;
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
 import ru.job4j.dreamjob.model.Candidate;
 import ru.job4j.dreamjob.model.Vacancy;
 import ru.job4j.dreamjob.repository.CandidateRepository;
@@ -20,24 +21,49 @@ public class SimpleCandidateService implements  CandidateService {
     @GuardedBy("this")
     private final CandidateRepository candidateRepository;
 
-    public SimpleCandidateService(CandidateRepository candidateRepository) {
+    private final FileService fileService;
+
+    public SimpleCandidateService(CandidateRepository candidateRepository, FileService fileService) {
         this.candidateRepository = candidateRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Candidate save(Candidate candidate) {
+    public Candidate save(Candidate candidate, FileDto image) {
+        saveNewFile(candidate, image);
         return candidateRepository.save(candidate);
+    }
+
+    private void saveNewFile(Candidate candidate, FileDto image) {
+        var file = fileService.save(image);
+        candidate.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return candidateRepository.deleteById(id);
+        var fileOptional = findById(id);
+        boolean result = false;
+        if (fileOptional.isPresent()) {
+            result = candidateRepository.deleteById(id);
+            fileService.deleteById(fileOptional.get().getFileId());
+        }
+        return result;
     }
 
     @Override
-    public boolean update(Candidate candidate) {
-        return candidateRepository.update(candidate);
+    public boolean update(Candidate candidate, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return candidateRepository.update(candidate);
+        }
+        /* если передан новый не пустой файл, то старый удаляем, а новый сохраняем */
+        var oldFileId = candidate.getFileId();
+        saveNewFile(candidate, image);
+        var isUpdated = candidateRepository.update(candidate);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
+
 
     @Override
     public Optional<Candidate> findById(int id) {
